Move pydantic models
